+++ src/main/java/lesson/http_jdbc/model/entity/CurrentConditionEntity.java
@@ -5,14 +5,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CurrentConditionEntity {
     private Long id;
-    private BigDecimal temp;
+    private Double temp;
     private Long cityId;
 }
